# Improve performance of `Char` generation