```
Benchmarks:                                                     BEFORE                AFTER
     44,360,879 randoms generated [System.Random Chars]      ~ 2,170 cycles/int    ~ 39.96 cycles/int
                    
  Next timing range-restricted System.Random.randomR:
     22,467,217 randoms generated [System.Random Chars]       ~ 1,949 cycles/int   ~ 78.89 cycles/int

Now 8 threads, reporting mean randoms-per-second-per-thread:
        568,199 randoms generated [System.Random Chars]       ~ 3,776 cycles/int  ~ 73.60 cycles/int
                    
        684,786 randoms generated [System.Random Chars]       ~ 3,133 cycles/int   ~ 295 cycles/int
```

Supersedes #74